suite du back, erreur de authoization

+++ Back-end/Taches/src/main/java/ToDo/Taches/repository/TacheRepository.java
@@ -4,7 +4,8 @@ import ToDo.Taches.model.Tache;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface TacheRepository extends JpaRepository<Tache, Integer> {
+public interface TacheRepository extends JpaRepository<Tache, Long> {
 
 }
